Fix the review's defects at their seams, and shape each package around one job
The review of runtime, cmd and samples reproduced ten defects and traced six of
them to three seams. One primitive, scheduler.Beside, now evaluates a function
beside its caller for spawn and every flow wrapper alike: guarded, derived from
the caller's evaluation rather than the run, inheriting one thread-local that
carries the run, the lane, the context, the attempt, the catching depth and the
name an update holds. Wait and a join's abort select on the caller's context,
so a timeout reaches a join. Fail is the one place that asks whether something
is catching, so an exhausted inner retry is one failed attempt of the outer.
Lock replaces the state plugin's mutex, waiting on the context and refusing
nesting through any thread the update started. Begin cancels the run's context
before waiting, closing a teardown window measured at 200 of 200 runs. Duration
refuses what no timer holds, the state copy consults its map only for
containers so a tuple round-trips, derivation refuses a signature it cannot
carry, and print reaches the host.

The tree is then shaped so each package does one job. The runtime's builtins
are a plugin, plugin/core, registered by import like any other, which removes
the facade's init and the scheduler's structural workaround for an import
cycle. The plugin registry is a type with a default instance and WithPlugins on
the compiler. scheduler.Evaluate owns the rules of what a run produced, leaving
artifact.Invoke a lookup. Reporter gains Printed, so a run tells its host one
thing. The facade states its rule for sentinels and names its compiler option
for what it configures.

Every former scheduler builtin test is rewritten through the exported surface.
A watcher test that now asserts found a release that could still cancel its
thread; the release waits for the watcher to exit.

+++ runtime/plugin/flow/report.go
@@ -14,19 +14,20 @@ import (
 // evaluation and ends the run. Guarding again would recover the same panic
 // twice and say nothing more.
 //
-// The lane is the caller's, not a new one: Detach inherits a thread's number
-// rather than taking one, so repeat(step, 3) is one entry on one lane advancing
-// 1, 2, 3 rather than three lanes.
+// The lane is the caller's, not a new one: an attempt runs on the lane that
+// made it, so repeat(3, step) is one entry on one lane advancing 1, 2, 3
+// rather than three lanes.
 //
 // Revisions:
 //   - 2026-09-20 01:40: initial creation
-func _Began(thread *starlark.Thread, name string, attempt int) {
+//   - 2026-09-21 08:09: takes the attempt in the reporter's own width
+func _Began(thread *starlark.Thread, name string, attempt int32) {
 	into := scheduler.Reporting(thread)
 	if into == nil {
 		return
 	}
 
-	into.Started(scheduler.Number(thread), name, int32(attempt))
+	into.Started(scheduler.Number(thread), name, attempt)
 }
 
 // _Finished tells whatever is watching how a wrapped function ended.
